+++ packages/docs/react-docs/qui-docs.config.ts
@@ -56,7 +56,7 @@ const navConfig: NavConfig[] = [
   {
     children: [
       {
-        id: "all-components",
+        id: "overview",
       },
       {
         group: "Data Display",
@@ -120,7 +120,11 @@ export default {
   knowledge: {
     global: {
       baseUrl: "https://react-next.qui.qualcomm.com",
-      exclude: ["**/installation+/**", "index.mdx"],
+      exclude: ["**/installation+/**", "index.mdx", "**/components+/overview*"],
+      exports: {
+        enabled: true,
+        exclude: ["**/components+/overview*"],
+      },
       extraFiles: [
         {
           contents: readFileSync(